dd
dd

+++ employee.js
@@ -30,36 +30,37 @@ $(document).ready(function(){
         $("#saveButton").click(function() {
             let names = document.getElementsByTagName("INPUT");
             //console.log(names[0].value);
-
-            for (let i = 0; i < names.length; i++) {        //no query since check doc exist
-                    db.collection("employees").doc(String.fromCharCode(i+65)).get().then(function(doc) {
-                        if (!doc.exists && names[i].value.replace(/\s+/g, '') !== '') {
-                            console.log('here');
-                            db.collection("employees").doc(String.fromCharCode(i+65)).set({
-                                name: names[i].value.replace(/\s+/g, ''),
-                                wage: 0
-                            });
-                        }
-                        else if (doc.exists && names[i].value.replace(/\s+/g, '') === '') {
-                            console.log('yellow');
-                            db.collection("employees").doc(String.fromCharCode(i+65)).delete();
-                        }
-                        else if (doc.exists && doc.data().name !== names[i].value.replace(/\s+/g, '')) {
-                            console.log('there');
-                            db.collection("employees").doc(String.fromCharCode(i+65)).set({
-                                name: names[i].value.replace(/\s+/g, ''),
-                                wage: 0
-                            });
-                        }
-                        else {
-                            console.log('yonder');
-                        }
-                    });
-                //}
-            }
-            window.alert("저장이 완료되었습니다");      //timeout of 5 seconds
+            let clicked = confirm("정보고 확실합니까? 이름이 변경됬거나 순서가 변경되는경우 전 정보가 지워질수도있습니다.");
+            if (clicked === true) {
+                for (let i = 0; i < names.length; i++) {        //no query since check doc exist
+                        db.collection("employees").doc(String.fromCharCode(i+65)).get().then(function(doc) {
+                            if (!doc.exists && names[i].value.replace(/\s+/g, '') !== '') {
+                                console.log('here');
+                                db.collection("employees").doc(String.fromCharCode(i+65)).set({
+                                    name: names[i].value.replace(/\s+/g, ''),
+                                    wage: 0
+                                });
+                            }
+                            else if (doc.exists && names[i].value.replace(/\s+/g, '') === '') {
+                                console.log('yellow');
+                                db.collection("employees").doc(String.fromCharCode(i+65)).delete();
+                            }
+                            else if (doc.exists && doc.data().name !== names[i].value.replace(/\s+/g, '')) {
+                                console.log('there');
+                                db.collection("employees").doc(String.fromCharCode(i+65)).set({
+                                    name: names[i].value.replace(/\s+/g, ''),
+                                    wage: 0
+                                });
+                            }
+                            else {
+                                console.log('yonder');
+                            }
+                        });
+                    //}
+                }
+                window.alert("저장이 완료되었습니다");      //timeout of 5 seconds
                 
-            
+            }
 
         });
     }
